fix: correctly export only types
